// src/Models/Login/logindetails.ts
export interface Logindetails {
    Flag: string;
    FlagValue: string;
    IsRefrenceCheck: string;
    IsOverPrintMandate: string;
    IsBulkMandate: string;
    IsMandate: string;
    IsMandateEdit: string;
    IsRefrenceEdit: string;
    IsEmandate: string;
    IsPhysical: string;
    IsZipShoreABPS: string;
    UserId: string;
    ReferenceId: string;
    UserName: string;
    Password: string;
    PasswordKey: string;
    UserCode: string;
    UserType: string;
    BranchId: string;
    BranchName: string;
    IsDefaultPswdChange: string;
    LastLogin: string;
    IsActive: string;
    IsDeleted: string;
    CreatedBy: string; 
    CreatedOn: string;
    AppId: string;
}